backend/requirements.txt
fastapi
uvicorn
torch
requests
pyannote.audio
faster-whisper
torchvision
torchtext
llama-cpp-python
cog
python-dotenv
python-multipart
pydub
python-magic
numpy
sqlalchemy
databases 
passlib 
cryptography
pycryptodome
bcrypt
PyMySQL